Medilens - Copy/backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
requests
google-generativeai
pytest
httpx
python-dotenv
